requirements.txt
future
emoji
lxml
wget
gtts
beautifulsoup4
requests==2.20.0
python-telegram-bot==12.8
sqlalchemy==1.3.23
psycopg2-binary
feedparser
hachoir
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
#googletrans
google_trans_new
jikanpy
speedtest-cli
coffeehouse==2.2.4
regex
bleach
git+https://github.com/trentm/python-markdown2.git
wikipedia
telethon==1.23.0
telegraph
heroku3
spamwatch
alphabet_detector
pyrate-limiter
cachetools
ujson
pretty_errors
tgCrypto>=1.2.2
Pyrogram==1.1.13
youtube_search_python
youtube_search
asyncio
dateparser==1.0.0
pymongo
dnspython
secureme
apscheduler
emoji-country-flag
countryinfo
html2text
bs4
bing_image_downloader
search_engine_parser
pytz 
redis
lyricsgenius
tswift
envparse
PyDictionary==2.0.1
#profanity (Credits to Julia)
better_profanity
textblob 
nudepy
motor
cloudscraper 
fontTools 
loguru 
glitch_this
NumPy
opencv-python-headless 
aiogram 
tzlocal==2.0
Markdown==3.3.6
